Day 3 part 2 golf attempt - 224 chars.
Instead of using std::set, use a 64 bit int (only 52 bits actually used) with a bit set for each char - 'a' sets bit 0, b sets bit 1 and so on. Then logical and the three lines after converting to the 64 bit representation, then increase the result by the number of the bit that has been set.